requirements.txt
asgiref==3.3.1
certifi==2021.5.30
chardet==4.0.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.6
django-crispy-forms==1.11.0
django-filter==21.1
django-import-export==2.5.0
et-xmlfile==1.0.1
Faker==8.14.0
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
numpy==1.20.1
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.1.0
playsound==1.3.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
standalone==1.0.1
tablib==3.0.0
text-unidecode==1.3
urllib3==1.26.5
xlrd==2.0.1
xlwt==1.3.0
